# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A lightweight local mock of the OpenAI REST API, covering Chat Completions\n(including streaming, tool calls, and multimodal content), Embeddings, Images,\nAudio, Models, Files, and Batch. Response shapes track the official OpenAI\nOpenAPI spec vendored at openai.yml; see memory.md for known deviations.",
        "title": "Mock OpenAI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/audio/speech": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates audio from the input text. Returns a fixed mock audio payload with a content-type matching the requested response_format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "audio/mpeg"
                ],
                "tags": [
                    "Audio"
                ],
                "summary": "Create speech",
                "parameters": [
                    {
                        "description": "Speech request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AudioSpeechRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    }
                }
            }
        },
        "/audio/transcriptions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transcribes audio into the input language. response_format controls the reply shape: json (default), text, srt, vtt, or verbose_json.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audio"
                ],
                "summary": "Create transcription",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Audio file to transcribe",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Model ID (e.g. whisper-1)",
                        "name": "model",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "json, text, srt, vtt, or verbose_json",
                        "name": "response_format",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AudioResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    }
                }
            }
        },
        "/audio/translations": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Translates audio into English. response_format controls the reply shape: json (default), text, srt, vtt, or verbose_json.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audio"
                ],
                "summary": "Create translation",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Audio file to translate",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Model ID (e.g. whisper-1)",
                        "name": "model",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "json, text, srt, vtt, or verbose_json",
                        "name": "response_format",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AudioResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    }
                }
            }
        },
        "/audio/voice_consents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audio"
                ],
                "summary": "List voice consents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.VoiceConsent"
                                    }
                                },
                                "object": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audio"
                ],
                "summary": "Create voice consent",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.VoiceConsent"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    }
                }
            }
        },
        "/audio/voice_consents/{consent_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audio"
                ],
                "summary": "Get voice consent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Voice consent ID",
                        "name": "consent_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.VoiceConsent"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audio"
                ],
                "summary": "Create voice consent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Voice consent ID (update variant only)",
                        "name": "consent_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.VoiceConsent"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audio"
                ],
                "summary": "Delete voice consent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Voice consent ID",
                        "name": "consent_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "deleted": {
                                    "type": "boolean"
                                },
                                "id": {
                                    "type": "string"
                                },
                                "object": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errs.Detail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "param": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "errs.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/errs.Detail"
                }
            }
        },
...
        "models.AudioResponse": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "models.AudioSpeechRequest": {
            "type": "object",
            "properties": {
                "input": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "response_format": {
                    "type": "string"
                },
                "speed": {
                    "type": "number"
                },
                "voice": {
                    "type": "string"
                }
            }
        },
...
        "models.VoiceConsent": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token auth. Any non-empty token is accepted unless API_KEY is configured (see README). Example: \"Bearer sk-mock\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}